feat!: relocate networkmap, standardize case and env cleanup
